gcp/code/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bmt-vm-runtime"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "google-cloud-storage>=2.16",
  "google-cloud-pubsub>=2.21",
  "whenever>=0.9.5",
]

[project.optional-dependencies]
vm = [
  "PyJWT>=2.0",
  "cryptography>=41.0",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["."]
include = ["lib*"]
